+mpfr, mpc, binutils, gmp and glibc

+++ gcc/build.sh
@@ -27,6 +27,16 @@ cd build
              --disable-fixincludes    \
              --with-system-zlib
 make -j$(nproc)
+ulimit -s -H unlimited
+chown -R tester .
+if ! (su tester -c "PATH=$PATH make -k check"); then
+	../contrib/test_summary -t
+	read -p 'Build failed tests. Proceed to installation anyway? [y/N] ' -n 1 -r < /dev/tty
+	echo
+	if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+		exit 1
+	fi
+fi
 sudo make install
 sudo su -c "ln -svr /usr/bin/cpp /usr/lib
 ln -sv gcc.1 /usr/share/man/man1/cc.1
